continue refactor *Maker to *Factory
