+++ tests/modelgauge_tests/test_dynamic_sut_finder.py
@@ -2,7 +2,7 @@ from unittest.mock import patch
 
 import pytest
 from modelgauge import dynamic_sut_finder
-from modelgauge.dynamic_sut_maker import UnknownProxyError
+from modelgauge.dynamic_sut_factory import UnknownProxyError
 from modelgauge.suts.huggingface_chat_completion import HuggingFaceChatCompletionServerlessSUT
 
 
@@ -11,7 +11,7 @@ def test_make_dynamic_sut():
         _ = dynamic_sut_finder.make_dynamic_sut_for("google:gemma:nebius:bogusproxy:20250101")
 
     with patch(
-        "modelgauge.suts.huggingface_sut_maker.HuggingFaceChatCompletionServerlessSUTMaker.find",
+        "modelgauge.suts.huggingface_sut_factory.HuggingFaceChatCompletionServerlessSUTFactory.find",
         return_value="cohere",
     ):
         registrable_sut = dynamic_sut_finder.make_dynamic_sut_for("google:gemma:cohere:hfrelay")
